refactor: routers.pyとschemas.pyを機能別モジュールに分割
- routers.py (1402行) を6つのモジュールに分割
  - categories.py, assets.py, snapshots.py, goals.py, dashboard.py, cash.py
- schemas.py (468行) を6つのモジュールに分割
  - category.py, asset.py, history.py, snapshot.py, goal.py, dashboard.py
- __init__.pyで後方互換性を維持
